Refactoring

+++ src/main/java/hantczak/studentDataStorage/grade/domain/GradeRepository.java
@@ -4,11 +4,18 @@ import java.util.List;
 
 public interface GradeRepository {
     List<Grade> getAllGrades();
+
     List<Grade> getStudentGrades(long studentId);
-    List<Grade> getAllGradesSorted(GradeSortType gradeSortType,int offset, int limit);
-    List<Grade> getAllStudentGradesSorted( long studentId,GradeSortType gradeSortType,int offset, int limit);
+
+    List<Grade> getAllGradesSorted(GradeSortType gradeSortType, long offset, long limit);
+
+    List<Grade> getAllStudentGradesSorted(long studentId, GradeSortType gradeSortType, long offset, long limit);
+
     Grade addGrade(Grade grade);
-    boolean updateGrade(Grade updatedGrade,long oldGradeId);
+
+    boolean updateGrade(Grade updatedGrade, long oldGradeId);
+
     boolean deleteGrade(long gradeToBeDeletedId);
+
     void deleteStudentGrades(long studentId);
 }
